--- src/api/mod.rs
@@ -1,8 +1,0 @@
-pub mod comment;
-pub mod cors;
-pub mod index;
-pub mod jwt;
-pub mod like;
-pub mod post;
-pub mod user;
-pub mod auth;
