+++ src/db/models/mod.rs
@@ -1,0 +1,4 @@
+pub mod user_models;
+pub mod comment_models;
+pub mod like_models;
+pub mod post_models;
