+++ src/main.rs
@@ -1,13 +1,8 @@
-pub mod api;
 pub mod db;
+pub mod routes;
 pub mod schema;
-use crate::api::cors::cors_fairing;
-use api::comment;
-use api::index::index;
-use api::like;
-use api::post;
-use api::user;
-use api::auth;
+use crate::routes::cors::cors_fairing;
+use routes::{auth_routes, comment_routes, index::index, like_routes, post_routes, user_routes};
 
 #[macro_use]
 extern crate rocket;
@@ -16,37 +11,34 @@ extern crate rocket;
 fn rocket() -> _ {
     let database_url = db::db_connection::DbPoolManager::get_database_url_from_dotenv();
     let pool = db::db_connection::DbPoolManager::init_pool(database_url.as_str());
-    rocket::build()
-        .manage(pool)
-        .attach(cors_fairing())
-        .mount(
-            "/api/v1",
-            routes![
-                index,
-                auth::user_new,
-                auth::login,
-                user::get_user,
-                user::user_all,
-                user::delete_user,
-                user::put_user,
-                post::get_post,
-                post::get_all_posts,
-                post::post_post,
-                post::put_post,
-                post::delete_post,
-                comment::get_comments,
-                comment::get_comments_user,
-                comment::delete_comment,
-                comment::put_comment,
-                comment::post_comment,
-                like::get_comment_likes,
-                like::comment_is_liked_by_user,
-                like::like_comment,
-                like::delete_comment_like,
-                like::get_post_likes,
-                like::post_is_liked_by_user,
-                like::like_post,
-                like::delete_post_like
-            ],
-        )
+    rocket::build().manage(pool).attach(cors_fairing()).mount(
+        "/api/v1",
+        routes![
+            index,
+            auth_routes::user_new,
+            auth_routes::login,
+            user_routes::get_user,
+            user_routes::user_all,
+            user_routes::delete_user,
+            user_routes::put_user,
+            post_routes::get_post,
+            post_routes::get_all_posts,
+            post_routes::post_post,
+            post_routes::put_post,
+            post_routes::delete_post,
+            comment_routes::get_comments,
+            comment_routes::get_comments_user,
+            comment_routes::delete_comment,
+            comment_routes::put_comment,
+            comment_routes::post_comment,
+            like_routes::get_comment_likes,
+            like_routes::comment_is_liked_by_user,
+            like_routes::like_comment,
+            like_routes::delete_comment_like,
+            like_routes::get_post_likes,
+            like_routes::post_is_liked_by_user,
+            like_routes::like_post,
+            like_routes::delete_post_like
+        ],
+    )
 }
