+++ src/routes/mod.rs
@@ -1,0 +1,8 @@
+pub mod comment_routes;
+pub mod cors;
+pub mod index;
+pub mod jwt;
+pub mod like_routes;
+pub mod post_routes;
+pub mod user_routes;
+pub mod auth_routes;
